LLTestProxy: an intermediate object that handles assertions
This stops us polluting RACSignal with additional methods. Also added semantics for enabling and disabling one-shot test proxies per object or globally. A per-Signal proxy can be useful for a more procedural style of testing, or tests that are not reliant on the repeated side effect semantics of a signal. This defaults to off, to encourage making non-repeatable side effects explicit