Removed test proxy idempotence & associated category on RACSignal
There was really no need for this to exist, it was an attempt to make
tests run faster and more predictably if they are operating on the same
events that the original Signal has sent.

We should really be concerned that a Signal behaves in a consistent
manner, and encourage the usage of Cold Signals. A Signal that does not
behave in a repeatable manner, after the first subscription will be
more likely to fail on successive subscriptions brought about by being
subscribed to in successive epectations.

If this behaviour is needed, in the case of a Signal that may only ever
be subscribed to once, they original signal should be multicasted &
replayed.